kvm: selftests: add ucall_test to test various ucall functionality
Tests will initialize the ucall implementation in 3 main ways,
depending on the test/architecture:

 1) by relying on a default ucall implementation being available
    without the need to do any additional setup
 2) by calling ucall_init() to initialize the default ucall
    implementation
 3) by using ucall_init_ops() to initialize a specific ucall
    implementation

and in each of these cases it may use the ucall implementation to
execute the standard ucall()/get_ucall() interfaces, or the new
ucall_shared()/get_ucall_shared() interfaces.

Implement a basic self-test to exercise ucall under all the scenarios
that are applicable for a particular architecture.

Signed-off-by: Michael Roth <michael.roth@amd.com>