x86: kvm: HACK: Allow testing of priv memfd approach
Add plumbing in KVM logic to allow private memfd series:
https://lore.kernel.org/linux-mm/20220310140911.50924-1-chao.p.peng@linux.intel.com/
to be tested with non-confidential VMs.

1) Existing hypercall KVM_HC_MAP_GPA_RANGE is modified to support
marking pages of the guest memory as privately accessed or
accessed in a shared fashion.

2) kvm_vcpu_is_private_gfn is overriden and defined to allow
guest accesses to be categorized as shared or private based on
the values set by KVM_HC_MAP_GPA_RANGE hypercall.

Signed-off-by: Vishal Annapurve <vannapurve@google.com>